src/main/kotlin/com/lfdev/bookstore/services/impl/AuthorServiceImpl.kt
package com.lfdev.bookstore.services.impl

import com.lfdev.bookstore.domain.author.AuthorEntity
import com.lfdev.bookstore.repositories.AuthorRepository
import com.lfdev.bookstore.services.AuthorService
import jakarta.transaction.Transactional
import org.springframework.data.repository.findByIdOrNull
import org.springframework.stereotype.Service

@Service
class AuthorServiceImpl(private val authorRepository: AuthorRepository): AuthorService {

    override fun create(authorEntity: AuthorEntity): AuthorEntity {
        require(null == authorEntity.id)
        return authorRepository.save(authorEntity)
    }

    override fun list(): List<AuthorEntity> {
        return authorRepository.findAll()
    }
    override fun getAuthor(id: Long): AuthorEntity? {
        return authorRepository.findByIdOrNull(id)
    }

    @Transactional
    override fun fullUpdate(id: Long, authorEntity: AuthorEntity): AuthorEntity {
        check(authorRepository.existsById(id))
        val normalizedAuthor = authorEntity.copy(id=id)
        return authorRepository.save(normalizedAuthor)
    }

}